+++ subsys/bluetooth/controller/ll_sw/nordic/lll/lll_internal.h
@@ -14,3 +14,4 @@ u32_t lll_evt_offset_get(struct evt_hdr *evt);
 u32_t lll_preempt_calc(struct evt_hdr *evt, u8_t ticker_id,
 		u32_t ticks_at_event);
 void lll_chan_set(u32_t chan);
+u8_t lll_entropy_get(u8_t len, void *rand);
